requirements.txt
langchain-core
langchain-community
langchain-chroma
langchain-google-genai
langchain-mistralai
google-generativeai
chromadb
tiktoken
python-dotenv
pypdf
flask
flask-cors
requests
